refactor: 버스트 판단 메서드 사용 위치 변경
