+++ src/main/java/blackjack/Blackjack.java
@@ -90,8 +90,4 @@ public class Blackjack {
         }
         return participantNameAndMatchResult;
     }
-
-    public boolean isBust(Player player) {
-        return player.isBust();
-    }
 }
