+++ src/test/java/blackjack/BlackjackTest.java
@@ -212,64 +212,4 @@ public class BlackjackTest {
             softly.assertThat(results.get("사나")).isEqualTo(MatchResult.DRAW);
         });
     }
-
-    @Test
-    void 해당_플레이어가_버스트_되었다면_TRUE를_반환한다() {
-        // given
-        Players players = new Players(List.of(
-                new Dealer(),
-                new Participant("시소")
-        ));
-
-        Player player = players.getPlayerByName("시소");
-
-        Deck deck = new Deck(new ArrayList<>(List.of(
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN)
-        )));
-        Blackjack blackjack = new Blackjack(players, deck);
-        blackjack.distributeInitialCards();
-        blackjack.addCard(player);
-        blackjack.addCard(player);
-        blackjack.addCard(player);
-
-        // when
-        boolean isBust = blackjack.isBust(player);
-
-        // then
-        Assertions.assertThat(isBust).isTrue();
-    }
-
-    @Test
-    void 해당_플레이어가_버스트_되었다면_FALSE를_반환한다() {
-        // given
-        Players players = new Players(List.of(
-                new Dealer(),
-                new Participant("시소")
-        ));
-
-        Player player = players.getPlayerByName("시소");
-
-        Deck deck = new Deck(new ArrayList<>(List.of(
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN),
-                new Card(CardSuit.SPADE, CardRank.TEN)
-        )));
-        Blackjack blackjack = new Blackjack(players, deck);
-        blackjack.distributeInitialCards();
-
-        // when
-        boolean isBust = blackjack.isBust(player);
-
-        // then
-        Assertions.assertThat(isBust).isFalse();
-    }
 }
