+++ src/test/java/player/PlayerTest.java
@@ -1,7 +1,13 @@
 package player;
 
+import blackjack.Blackjack;
+import card.Card;
+import card.CardRank;
+import card.CardSuit;
 import card.Deck;
 import card.DeckGenerator;
+import java.util.ArrayList;
+import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 
@@ -43,5 +49,65 @@ public class PlayerTest {
         // when & then
         Assertions.assertThat(player)
                 .isEqualTo(new Participant(targetName));
+    }
+
+    @Test
+    void 해당_플레이어가_버스트_되었다면_TRUE를_반환한다() {
+        // given
+        Players players = new Players(List.of(
+                new Dealer(),
+                new Participant("시소")
+        ));
+
+        Player player = players.getPlayerByName("시소");
+
+        Deck deck = new Deck(new ArrayList<>(List.of(
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN)
+        )));
+        Blackjack blackjack = new Blackjack(players, deck);
+        blackjack.distributeInitialCards();
+        blackjack.addCard(player);
+        blackjack.addCard(player);
+        blackjack.addCard(player);
+
+        // when
+        boolean isBust = player.isBust();
+
+        // then
+        Assertions.assertThat(isBust).isTrue();
+    }
+
+    @Test
+    void 해당_플레이어가_버스트_되었다면_FALSE를_반환한다() {
+        // given
+        Players players = new Players(List.of(
+                new Dealer(),
+                new Participant("시소")
+        ));
+
+        Player player = players.getPlayerByName("시소");
+
+        Deck deck = new Deck(new ArrayList<>(List.of(
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN),
+                new Card(CardSuit.SPADE, CardRank.TEN)
+        )));
+        Blackjack blackjack = new Blackjack(players, deck);
+        blackjack.distributeInitialCards();
+
+        // when
+        boolean isBust = player.isBust();
+
+        // then
+        Assertions.assertThat(isBust).isFalse();
     }
 }
